// extension/payload.types.d.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: payloads.go

export type WebsocketPayload = IdentifyPayload | IdentifyPayload | IdentifiedPayload

export interface IdentifyPayload {
  kind: 'IDENTIFY';
}
export interface IdentityPayload {
  kind: 'IDENTITY';
  user_token: string /* uuid */;
}
export interface IdentifiedPayload {
  kind: 'IDENTIFIED';
}